Minify CSS and XHTML

+++ build.sh
@@ -3,10 +3,21 @@
 set -euxo pipefail
 
 cargo run
+cat index.xhtml | minify --mime='application/xhtml+xml' > index.min
+mv index.{min,xhtml}
 
-cp res/style/{source,minified}.css
-runhaskell downscale.hs 80 140 >> res/style/minified.css
+{ cat res/style/source.css
+  runhaskell downscale.hs 80 140
+} | minify --mime='text/css' > res/style/minified.css
 
-rm -rf res/images/low_res
-mkdir res/images/low_res
-mogrify -resize 'x555>' -quality 88 -format webp -auto-orient -path res/images/{low_res,hi_res/*}
+new=($(cd res/images/hi_res; cat $(ls -1 | sort) | sha1sum))
+old=$(cat res/images/low_res/originals.sha1 || true)
+if [[ "$new" != "$old" ]]; then
+    rm -rf res/images/low_res
+    mkdir res/images/low_res
+    mogrify \
+        -resize 'x555>' -quality 88 \
+        -format webp -auto-orient \
+        -path res/images/{low_res,hi_res/*}
+    echo -n "$new" > res/images/low_res/originals.sha1
+fi
